# Social-Distribtion-Engine-MelloxAI-SDE/specs/001-social-sde/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MelloxAI Social Distribution Engine (SDE)
  description: |
    Backend service for publishing and scheduling social media posts.

    ## Authentication
    All endpoints except `/healthz` require:
    - `Authorization: Bearer <SDE_API_TOKEN>`
    - `X-Signature: hex(hmac_sha256(SDE_SIGNING_SECRET, raw_body))`
    - `X-Timestamp`: Unix timestamp, reject if |now - ts| > 300s

    ## Error Response Format
    All errors follow this structure:
    ```json
    {
      "error_code": "ERROR_CODE",
      "detail": "Human-readable explanation",
      "how_to_fix": "Actionable next step",
      "field": "optional field name",
      "platform": "optional platform"
    }
    ```

    ## Key Concepts
    - **Post**: Logical publish request
    - **Target**: Individual platform publish attempt
    - **Idempotency**: Same idempotency_key returns existing post
  version: 1.0.0
  contact:
    name: MelloxAI Engineering
    url: https://mellox.ai
servers:
  - url: https://sde.mellox.ai/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /publish:
    post:
      summary: Publish immediately
      description: |
        Submit a post for immediate publishing. Sets scheduled_at to now().
        Returns 201 if new, 200 if idempotency key already exists.
      operationId: publishNow
      tags:
        - Publishing
      security:
        - BearerAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: Idempotent duplicate - existing post returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "201":
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Platform-specific validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /schedule:
    post:
      summary: Schedule for later
      description: |
        Submit a post for scheduled publishing. scheduled_at must be in the future.
        Returns 201 if new, 200 if idempotency key already exists.
      operationId: schedulePost
      tags:
        - Publishing
      security:
        - BearerAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: Idempotent duplicate - existing post returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "201":
          description: Post scheduled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          description: Validation error (scheduled_at in past)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Platform-specific validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    TimestampAuth:
      type: apiKey
      in: header
      name: X-Timestamp
      description: Unix timestamp (reject if > 5 min old)
  schemas:
...
    PublishRequest:
      type: object
      required:
        - workspace_id
        - brand_id
        - idempotency_key
        - scheduled_at
        - targets
      properties:
        workspace_id:
          type: string
          description: Workspace identifier
        brand_id:
          type: string
          description: Brand/tenant identifier
        idempotency_key:
          type: string
          description: Unique key for idempotency (generated by MelloxAI)
        scheduled_at:
          type: string
          format: date-time
          description: Publish time (now() for immediate, future for scheduled)
        timezone:
          type: string
          description: Timezone for scheduled_at interpretation
        targets:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PublishTarget"
    PublishTarget:
      type: object
      required:
        - account_id
        - platform
        - content
      properties:
        account_id:
          type: string
          format: uuid
          description: Connected account UUID
        platform:
          type: string
          enum: [x, linkedin, facebook, dryrun]
        content:
          $ref: "#/components/schemas/TargetContent"
    TargetContent:
      type: object
      required:
        - text
      properties:
        text:
          type: string
        media:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [image, video]
              url:
                type: string
                format: uri
              alt:
                type: string
        link:
          type: string
          format: uri
        first_comment:
          type: string
          nullable: true

    # Response Schemas
    PublishResponse:
      type: object
      properties:
        post_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, PARTIAL, COMPLETED, FAILED, CANCELLED]
        idempotency_duplicate:
          type: boolean
          description: True if this was an idempotent duplicate
        targets:
          type: array
          items:
            $ref: "#/components/schemas/TargetStatus"
    TargetStatus:
      type: object
      properties:
        target_id:
          type: string
          format: uuid
        platform:
          type: string
        status:
          type: string
          enum: [PENDING, QUEUED, PUBLISHING, PUBLISHED, RETRYING, FAILED, CANCELLED]
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        detail:
          type: string
        how_to_fix:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        error_code:
          type: string
          const: VALIDATION_ERROR
        detail:
          type: string
        how_to_fix:
          type: string
        field_errors:
          type: array
          items:
            type: object
            properties:
              platform:
                type: string
              field:
                type: string
              error:
                type: string
...